requirements.txt
websockets
pyzmq
python-bitcoinlib